Update namespace for the cellcap package
