+++ build_docker.sh
@@ -2,4 +2,4 @@
 
 VERSION=$(cat VERSION)
 
-docker build --file docker/Dockerfile . --tag perturbations:${VERSION}
+docker build --file docker/Dockerfile . --tag cellcap:${VERSION}
