+++ cellcap/__init__.py
@@ -1,0 +1,10 @@
+"""Namespace for the package"""
+
+from .module_pytorch import CellCap
+from .utils import generate_simulated_dataset
+
+
+__all__ = [
+    CellCap,
+    generate_simulated_dataset,
+]
