+++ cellcap/utils.py
@@ -1,6 +1,34 @@
 """Utility functions"""
 
 import torch
+import anndata
+import numpy as np
+
+from scvi.data import synthetic_iid
+
+
+def _random_one_hot(n_classes: int, n_samples: int):
+    # https://stackoverflow.com/questions/45093615/random-one-hot-matrix-in-numpy
+    return np.eye(n_classes)[np.random.choice(n_classes, n_samples)]
+
+
+def generate_simulated_dataset() -> anndata.AnnData:
+
+    # random data via scvi-tools
+    adata = synthetic_iid()
+    n = adata.shape[0]
+    del adata.obsm['protein_expression']
+
+    # add in necessary fields
+    obsm_data = {'cond': _random_one_hot(2, n),
+                 'cont': _random_one_hot(2, n),
+                 'target': _random_one_hot(2, n),
+                 'donor': _random_one_hot(2, n)}
+    for k, v in obsm_data.items():
+        adata.obsm[k] = v
+    adata.obs['pert'] = _random_one_hot(2, n)[:, 0]
+
+    return adata
 
 
 class Metrics:
